clean up age, updated_at (#61)
This commit cleans up the age and updated_at fields. It now uses dates
that are relevant to now instead of hours (i.e. "six days ago", "four
hours ago", "one month ago"). It also uses the correct dates now, as age
used to use the time between the created_at date and the updated_at
date. It now sorts the dates correctly as well.

Unrelated, it adds `go get` to the readme, which is required to build
the software.

Fixes: https://github.com/minio/buzz/issues/57