merge reviewers, review activity into colored tags (#67)
This commit merges the reviewers and review activity sections of the
table. Now, it lists all users who have been requested for a review or
have begun a review voluntarily. The status of their review is visible
in the color of their tag. Pending for yellow, accepted for green, or
requested changes for red.

Fixes: https://github.com/minio/buzz/issues/58