reviewers: add commented, but not reviewing users
This commit adds users who comment on a PR line of code but do not
actually review the code to the reviewers list. This commit adds them as
a gray tag, so that they can be distinguished from those who are
actually reviewing. Gray now means that they are expressing interest in
a change but not necessarily reviewing the code.

This does not include the actual PR author, who will often comment on
their own code in a reply to someone else.

Fixes: https://github.com/minio/buzz/issues/70