add commented reviewers, remove global variables (#72)
* reviewers: add commented, but not reviewing users

This commit adds users who comment on a PR line of code but do not
actually review the code to the reviewers list. This commit adds them as
a gray tag, so that they can be distinguished from those who are
actually reviewing. Gray now means that they are expressing interest in
a change but not necessarily reviewing the code.

This does not include the actual PR author, who will often comment on
their own code in a reply to someone else.

Fixes: https://github.com/minio/buzz/issues/70

* eliminate some global variables, clean code

This commit eliminates global variables related to retrieving issues and
pull requests. I'm not sure why they were there in the first place, as
they'd cause issues if more than one request was active on the server at
any point in time, and they're not idiomatic. This commit additionally
reduces code in the same files because it wasn't necessary. Instead, it
opts to use the Go-GitHub library, which allows for easy use of the API
instead of retrieving the URLs on our own.

Fixes: https://github.com/minio/buzz/issues/71
Helps: https://github.com/minio/buzz/issues/66
Helps: https://github.com/minio/buzz/issues/59
Helps: https://github.com/minio/buzz/issues/23